Move ServerDeclarationView to control package
